feat(routes): migrate /blog to /garden terminology (Phase 1)
Grove terminology migration Phase 1 - rename blog routes to garden:

- Add 301 redirects for /blog → /garden (SEO preservation)
- Rename routes/blog/ → routes/garden/
- Rename routes/admin/blog/ → routes/admin/garden/
- Update navigation links in tenant-nav.ts, defaults.ts
- Update admin sidebar and dashboard links
- Update RSS feed URLs to /garden/
- Update cache keys from blog: to garden: prefix
- Update InternalsPostViewer component link

Phase 2 (admin → arbor) will follow in a separate PR.
